# docs/swagger.yaml
definitions:
...
  handlers.updateData:
    properties:
      channel_id:
        type: string
      description:
        type: string
    type: object
...
  models.DiscordMessage:
    properties:
      template:
        type: string
    type: object
...
  models.Project:
    properties:
      added_by:
        type: string
      channel_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      project_name:
        type: string
      project_source_auth_token:
        type: string
      project_source_url:
        type: string
      updated_at:
        type: string
      webhook_origin:
        type: string
      webhook_secret:
        type: string
      webhook_url:
        type: string
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      error:
        items:
          type: string
        type: array
      status:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/projects:
    get:
      consumes:
      - application/json
      description: Retrieves an array of all existing projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Project'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project from the request body
      parameters:
      - description: Project to create
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Create a new project
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a project using the provided ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Delete a project by ID
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Retrieves a single project using the provided ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get a project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates a project's details using the provided ID and request body
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.updateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update a project by ID
      tags:
      - projects
  /api/v1/projects/{id}/send-message:
    post:
      consumes:
      - application/json
      description: Sends a message to the Discord channel associated with the specified
        project ID. The message is rendered using a Go template provided in the request
        body.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Message payload with template
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DiscordMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Send a message to a Discord channel
      tags:
      - projects
  /api/v1/projects/{id}/source:
    put:
      consumes:
      - application/json
      description: Updates a project's source authentication token using the provided
        ID and request body
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  items:
                    type: string
                  type: array
              type: object
      summary: Update a project's source authentication token by ID
      tags:
      - projects
...
swagger: "2.0"